API, getnc returns dumb_ssize_t instead of size_t
This is necessary because getnc returns -1 on error. Spec updated with
the new function signatures.

Fixed resulting comparisons between signed and unsigned that gcc 7.2
shows when we compile with -Wall -Wextra -pedantic.

I have not compiled with -Wconversion to find implicit integer coercion
from assigning to narrower types, that would find far too many intended
coercions in the music rendering.