Check whether CFLAGS are supported before adding
Add non-essential CFLAGS only if the compiler supports it.
Enable compiling on MSVC and Clang without error-msgs or warnings
due to non-existing parameters. Note that MSVC has not been tested
beyond the fact that it compiles.

This also makes compiling on non-x86/x86_64 architectures work in the
default configuration, by using the existance of the -msse parameter as
an indication whether SSE is supported. This may not be the best check,
but is still better than requiring manual configuration.